Played around with ada and vending variables as lgl to see if it changed the results of my calculations - it did not.
Initially I was unable to knit but was able to run code in the console.  After now being able to knit, I have changed the written text to update my calculations.

Also, I reconsidered my response to the proportion of distinct station entrance/exits without vending machines that allow entrance by changing the denominator to the number of distinct stations.